docs: refactor "templated_lsig" example to be exemplary for v1
- move demo code to its own file
