+++ examples/templated_lsig/test_demo.py
@@ -1,10 +1,11 @@
 from tests.conftest import check_application_artifacts_output_stability
 
-from examples.templated_lsig.main import app, demo
+from examples.templated_lsig import demo
+from examples.templated_lsig.main import app
 
 
 def test_demo() -> None:
-    demo()
+    demo.main()
 
 
 def test_output_stability() -> None:
